# server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "乐清商会 API",
    "version": "1.0.0",
    "description": "乐清商会后端 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    },
    {
      "url": "https://yueqing-chamber.vercel.app",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Member": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "company": { "type": "string" },
          "position": { "type": "string" },
          "phone": { "type": "string" },
          "industry": { "type": "string" }
        }
      },
      "Activity": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "date": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "maxParticipants": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["认证"],
        "summary": "用户注册",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string" },
                  "password": { "type": "string" },
                  "email": { "type": "string" }
                },
                "required": ["phone", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "注册成功" },
          "400": { "description": "请求参数错误" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["phone", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/members": {
      "get": {
        "tags": ["会员管理"],
        "summary": "获取会员列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": { "description": "成功返回会员列表" }
        }
      }
    },
    "/api/members/{id}/details": {
      "get": {
        "tags": ["会员管理"],
        "summary": "获取会员详情",
        "description": "返回会员完整信息、最近活动和报名记录",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "会员ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回会员详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "phone": { "type": "string" },
                        "email": { "type": "string", "nullable": true },
                        "company": { "type": "string" },
                        "position": { "type": "string", "nullable": true },
                        "joindate": { "type": "string", "format": "date-time", "nullable": true },
                        "status": { "type": "string" },
                        "recentActivities": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "activityId": { "type": "string" },
                              "activityTitle": { "type": "string" },
                              "activityDate": { "type": "string", "format": "date-time" },
                              "activityLocation": { "type": "string" },
                              "status": { "type": "string" },
                              "registeredAt": { "type": "string", "format": "date-time" }
                            }
                          }
                        },
                        "registrationCount": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "会员不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/activities": {
      "get": {
        "tags": ["活动管理"],
        "summary": "获取活动列表",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": { "description": "成功返回活动列表" }
        }
      }
    }
  }
}